Refactoring oriantation handling.
